Lab 6 fork completed

+++ lab6/include/cpio.h
@@ -1,4 +1,4 @@
-void list(unsigned int addr);
-void print_content(char *file, unsigned int addr);
+void list(uint64_t addr);
+void print_content(char *file, uint64_t addr);
 void exec(char *file_name,char *const argv[]);
 void copy_content(char *file,void **addr, uint64_t *size);
