+++ lab6/src/cpio.c
@@ -7,7 +7,7 @@
 #include "scheduler.h"
 #include "task.h"
 #include "mmu.h"
-extern unsigned int cpio_start;
+extern uint64_t cpio_start;
 
 typedef struct cpio_newc_header {  //cpio new ascii struct
 		   char	   c_magic[6];
@@ -27,7 +27,7 @@ typedef struct cpio_newc_header {  //cpio new ascii struct
 }CPIO_H ;
 
 
-void list(uint32 addr){ // proceed ls
+void list(uint64_t addr){ // proceed ls
     CPIO_H *cpio=(CPIO_H*) addr;
     while(strncmp(cpio->c_magic,"070701\0",6)==1){ //c_magic is always "070701"
         int namesize=a16ntoi(cpio->c_namesize, 8);
@@ -53,7 +53,7 @@ void list(uint32 addr){ // proceed ls
     }
 }
 
-uint32 find_file_addr(char *file_name, uint32 addr){
+uint64_t find_file_addr(char *file_name, uint64_t addr){
     CPIO_H *cpio=(CPIO_H*) addr;
     while(strncmp(cpio->c_magic,"070701\0",6)==1){
         int namesize=a16ntoi(cpio->c_namesize, 8);
@@ -80,7 +80,7 @@ uint32 find_file_addr(char *file_name, uint32 addr){
     return NULL;
 }
 
-void* getContent(uint32 addr,int *length){
+void* getContent(uint64_t addr,int *length){
     CPIO_H *cpio=(CPIO_H*) addr;
     int namesize=a16ntoi(cpio->c_namesize, 8);
     int filesize=a16ntoi(cpio->c_filesize, 8);
@@ -89,8 +89,8 @@ void* getContent(uint32 addr,int *length){
     return addr;
 }
 
-void print_content(char *file, uint32 addr){
-    uint32 f_addr=find_file_addr(file,addr);
+void print_content(char *file, uint64_t addr){
+    uint64_t f_addr=find_file_addr(file,addr);
     if(f_addr != NULL){
         int length;
         char *content_addr = getContent(f_addr, &length);
@@ -106,10 +106,10 @@ void print_content(char *file, uint32 addr){
 }
 
 void copy_content(char *file,void **addr, uint64_t *size){
-    uint32 f_addr=find_file_addr(file,cpio_start);
+    uint64_t f_addr=find_file_addr(file,cpio_start);
     if(f_addr != NULL){
         int length;
-        uint32 address = getContent(f_addr, &length);
+        uint64_t address = getContent(f_addr, &length);
         char *content_addr = address;
         if(address%4 != 0){
             address /= 4;
@@ -132,10 +132,10 @@ void copy_content(char *file,void **addr, uint64_t *size){
 }
 
 void exec(char *file,char *const argv[]){
-    uint32 f_addr=find_file_addr(file,cpio_start);
+    uint64_t f_addr=find_file_addr(file,cpio_start);
     if(f_addr != NULL){
         int length;
-        uint32 address = getContent(f_addr, &length);
+        uint64_t address = getContent(f_addr, &length);
         char *content_addr = address;
         if(address%4 != 0){
             address /= 4;
